requirements.txt
streamlit
google-generativeai
sentence-transformers
faiss-cpu
pandas
numpy
filelock
pydantic
python-dateutil
langgraph
python-dotenv

